@@ -6,22 +6,23 @@
   "dependencies": {
     "@apollo/react-hooks": "^4.0.0",
     "apollo-boost": "^0.4.9",
-    "colorthief": "^2.3.0",
-    "graphql": "^14.5.8",
-    "react": "^16.10.2",
-    "react-dom": "^16.10.2",
+    "colorthief": "^2.6.0",
+    "graphql": "^14.7.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-easy-emoji": "^1.8.1",
-    "react-headroom": "^3.0.0",
-    "react-reveal": "^1.2.2",
-    "react-scripts": "3.2.0"
+    "react-headroom": "^3.2.1",
+    "react-reveal": "^1.2.2"
   },
   "scripts": {
+    "dev": "vite",
     "predeploy": "npm run build",
     "start": "react-scripts start",
-    "build": "react-scripts build",
+    "build": "vite build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "deploy": "netlify deploy --prod"
+    "deploy": "netlify deploy --prod",
+    "preview": "vite preview"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -38,5 +39,8 @@
       "last 1 safari version"
     ]
   },
-  "devDependencies": {}
+  "devDependencies": {
+    "@vitejs/plugin-react": "^4.7.0",
+    "vite": "^6.4.1"
+  }
 }
