@@ -4,6 +4,7 @@ import SoftwareSkill from "../../components/softwareSkills/SoftwareSkill";
 import { skillsSection } from "../../portfolio";
 import { Fade } from "react-reveal";
 import StyleContext from "../../contexts/StyleContext";
+import developerActivity from "../../assets/images/developerActivity.svg";
 
 export default function Skills() {
   const { isDark } = useContext(StyleContext);
@@ -17,7 +18,7 @@ export default function Skills() {
           <div className="skills-image-div">
             <img
               alt="Bhavay Working"
-              src={require("../../assets/images/developerActivity.svg")}
+              src={developerActivity}
             ></img>
           </div>
         </Fade>
